fix: 보이스 문제 정답 수정

+++ backend/homework/src/main/java/com/eefy/homework/domain/homework/service/HomeworkServiceImpl.java
@@ -416,7 +416,7 @@ public class HomeworkServiceImpl implements HomeworkService {
         String answer) {
         String score = aiServiceFeignClient.getAnnounceScore(
             new EvaluateAnnounceRequest(homeworkStudentQuestion.getFilePath(), answer));
-        homeworkStudentQuestion.updateScore(Integer.parseInt(score.substring(1, 4)) * 20);
+        homeworkStudentQuestion.updateScore((int) (Double.parseDouble(score.substring(1, 4)) * 20));
     }
 
 }
